requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
